+++ packages/plugin-bm-api/src/graphql/schema/element.ts
@@ -1,4 +1,4 @@
-import gql from 'graphql-tag';
+import gql from "graphql-tag";
 
 export const types = () => `
 
@@ -20,6 +20,7 @@ export const types = () => `
     _id: String!
     name: String
     quick: Boolean
+    visibleName: Boolean
     icon: String
     content: String
     note: String
@@ -48,7 +49,7 @@ export const types = () => `
 `;
 
 export const queries = `
-  bmElements(branchId:String, categories: [String],name: String, page:Int, perPage:Int,quick: Boolean,sortField:String, sortDirection:Int): ListElement
+  bmElements(branchId:String, categories: [String],name: String, page:Int, perPage:Int,quick: Boolean,sortField:String, sortDirection:Int,icon:Boolean): ListElement
   bmElementDetail(_id:String!): Element
   bmElementCategoryies(parentId:String): [ElementCategory]
   bmElementsInit: JSON
@@ -67,7 +68,9 @@ const params = `
   itineraryId: String,
   location: BMSLocationInput,
   quick: Boolean,
-  branchId: String
+  branchId: String,
+  icon: String,
+  visibleName: Boolean
 `;
 
 export const mutations = `