+++ packages/plugin-sales-api/src/graphql/resolvers/mutations/utils.ts
@@ -911,13 +911,15 @@ export const doScoreCampaign = async (
       const scoreCampaignTypes = (pipeline?.paymentTypes || []).filter(
         ({ scoreCampaignId }) => !!scoreCampaignId
       );
-
-      target.exludeAmount = Object.entries(doc.paymentsData)
-        .filter(([type]) => !scoreCampaignTypes.includes(type))
+      target.excludeAmount = Object.entries(doc.paymentsData)
+        .filter(
+          ([type]) => !scoreCampaignTypes.map(({ type }) => type).includes(type)
+        )
         .map(([type, obj]) => ({
           type,
           ...obj,
-        }));
+        }))
+        .reduce((sum, payment) => sum + (payment?.amount || 0), 0);
       for (const type of types) {
         const paymentType = scoreCampaignTypes.find(
           (paymentType) => paymentType.type === type
@@ -987,4 +989,35 @@ export const doScoreCampaign = async (
       }
     }
   }
+};
+
+export const confirmLoyalties = async (
+  subdomain: string,
+  _id: string,
+  deal: IDeal
+) => {
+  const confirmItems = deal.productsData || [];
+
+  if (!confirmItems.length) {
+    return;
+  }
+
+  const [customerId] = (await getCustomerIds(subdomain, "deal", _id)) || [];
+
+  try {
+    await sendLoyaltiesMessage({
+      subdomain,
+      action: "confirmLoyalties",
+      data: {
+        checkInfo: {},
+        extraInfo: {
+          ...(deal.extraData || {}),
+          ownerType: "customer",
+          ownerId: customerId || null,
+          targetType: "sales",
+          targetId: _id,
+        },
+      },
+    });
+  } catch (e) {}
 };