+++ packages/plugin-sales-api/src/graphql/schema/deal.ts
@@ -55,6 +55,7 @@ export const types = ({ contacts, clientPortal }) => `
   input SalesProductField {
     productId : String
     quantity: Int
+    unitPrice: Float
   }
 
 `;
@@ -62,6 +63,7 @@ export const types = ({ contacts, clientPortal }) => `
 const dealMutationParams = `
   paymentsData: JSON,
   productsData: JSON,
+  extraData: JSON,
 `;
 
 const commonQueryParams = `
@@ -134,7 +136,7 @@ const archivedDealsParams = `
 
 export const queries = `
   dealDetail(_id: String!, clientPortalCard: Boolean): Deal
-  checkDiscount(_id: String!,products:[SalesProductField]):JSON
+  checkDiscount(_id: String!,products:[SalesProductField], couponCode: String):JSON
   deals(${listQueryParams}): [DealListItem]
   dealsTotalCount(${listQueryParams}): Int
   archivedDeals(