+++ packages/plugin-tasks-ui/src/automations/automation.tsx
@@ -2,11 +2,14 @@ import React from 'react';
 import BoardItemForm from './containers/BoardItemForm';
 import SelectBoard from './components/SelectBoard';
 import ActionResult from './components/ActionResult';
+import TriggerForm from './components/TriggerForm';
 
-const Automations = props => {
+const Automations = (props) => {
   const { componentType, target } = props;
 
   switch (componentType) {
+    case 'triggerForm':
+      return <TriggerForm {...props} />;
     case 'actionForm':
       return <BoardItemForm {...props} />;