+++ packages/ui-inbox/src/settings/integrations/graphql/queries.ts
@@ -45,6 +45,8 @@ const integrationDetail = `
       callData {
         header
         description
+        secondPageHeader
+        secondPageDescription
         departments {
           _id
           name