+++ packages/ui-sales/src/deals/components/product/ProductForm.tsx
@@ -1,35 +1,42 @@
-import { Add, FlexRowGap, FooterInfo, FormContainer } from "../../styles";
-import lodash from "lodash";
-import { Alert, __ } from "@erxes/ui/src/utils";
+import { gql } from '@apollo/client';
+import SelectCompanies from "@erxes/ui-contacts/src/companies/containers/SelectCompanies";
+import ProductCategoryChooser from "@erxes/ui-products/src/components/ProductCategoryChooser";
+import ProductChooser from "@erxes/ui-products/src/containers/ProductChooser";
+import { IProduct, IProductCategory } from "@erxes/ui-products/src/types";
+import SelectTags from "@erxes/ui-tags/src/containers/SelectTags";
+import client from '@erxes/ui/src/apolloClient';
+import { IUser } from "@erxes/ui/src/auth/types";
+import SelectBrands from "@erxes/ui/src/brands/containers/SelectBrands";
 import {
   ControlLabel,
   FormGroup,
   ModalTrigger,
   Table,
 } from "@erxes/ui/src/components";
-import { IDeal, IPaymentsData, IProductData } from "../../types";
+import Button from "@erxes/ui/src/components/Button";
+import EmptyState from "@erxes/ui/src/components/EmptyState";
+import FormControl from "@erxes/ui/src/components/form/Control";
+import Icon from "@erxes/ui/src/components/Icon";
 import { TabTitle, Tabs } from "@erxes/ui/src/components/tabs";
-
-import Button from '@erxes/ui/src/components/Button';
-import EmptyState from '@erxes/ui/src/components/EmptyState';
-import FormControl from '@erxes/ui/src/components/form/Control';
-import { IProduct } from '@erxes/ui-products/src/types';
-import { IProductCategory } from '@erxes/ui-products/src/types';
-import Icon from '@erxes/ui/src/components/Icon';
-import { ModalFooter } from '@erxes/ui/src/styles/main';
-import PaymentForm from './PaymentForm';
-import ProductCategoryChooser from '@erxes/ui-products/src/components/ProductCategoryChooser';
-import ProductChooser from '@erxes/ui-products/src/containers/ProductChooser';
-import ProductItem from '../../containers/product/ProductItem';
-import ProductTotal from './ProductTotal';
-import React from 'react';
-import SelectBranches from '@erxes/ui/src/team/containers/SelectBranches';
-import SelectBrands from '@erxes/ui/src/brands/containers/SelectBrands';
-import SelectCompanies from '@erxes/ui-contacts/src/companies/containers/SelectCompanies';
-import SelectDepartments from '@erxes/ui/src/team/containers/SelectDepartments';
-import SelectTags from "@erxes/ui-tags/src/containers/SelectTags";
-import styled from 'styled-components';
-import { IUser } from '@erxes/ui/src/auth/types';
+import { ModalFooter } from "@erxes/ui/src/styles/main";
+import SelectBranches from "@erxes/ui/src/team/containers/SelectBranches";
+import SelectDepartments from "@erxes/ui/src/team/containers/SelectDepartments";
+import { Alert, __ } from "@erxes/ui/src/utils";
+import { isEnabled } from '@erxes/ui/src/utils/core';
+import lodash from "lodash";
+import React from "react";
+import styled from "styled-components";
+import ProductItem from "../../containers/product/ProductItem";
+import { queries } from '../../graphql';
+import { Add, FlexRowGap, FooterInfo, FormContainer } from "../../styles";
+import {
+  IDeal,
+  IPaymentsData,
+  IProductData,
+  dealsProductDataMutationParams,
+} from "../../types";
+import PaymentForm from "./PaymentForm";
+import ProductTotal from "./ProductTotal";
 
 const TableWrapper = styled.div`
   overflow: auto;
@@ -65,6 +72,8 @@ type Props = {
   onChangeProductsData: (productsData: IProductData[]) => void;
   saveProductsData: () => void;
   onChangePaymentsData: (paymentsData: IPaymentsData) => void;
+  dealsCreateProductData: (variables: dealsProductDataMutationParams) => void;
+  onChangeExtraData: (extraData: any) => void;
   productsData: IProductData[];
   products: IProduct[];
   paymentsData?: IPaymentsData;
@@ -90,6 +99,7 @@ type State = {
   tempId: string;
   filterValues: any;
   advancedView?: boolean;
+  couponCode?: string;
 };
 
 class ProductForm extends React.Component<Props, State> {
@@ -117,20 +127,36 @@ class ProductForm extends React.Component<Props, State> {
   }
 
   duplicateProductItem = (_id) => {
-    const { productsData, onChangeProductsData } = this.props;
+    const {
+      productsData,
+      onChangeProductsData,
+      dealsCreateProductData,
+      dealQuery,
+    } = this.props;
 
     const productData: any = productsData.find((p) => p._id === _id);
 
-    productsData.push({
+    const docs: any[] = [];
+
+    const data = {
       ...productData,
       _id: Math.random().toString(),
-    });
+    };
+
+    docs.push(data);
+    productsData.push(data);
 
     onChangeProductsData(productsData);
 
     for (const productData of productsData) {
       this.calculatePerProductAmount("discount", productData);
     }
+
+    dealsCreateProductData({
+      proccessId: localStorage.getItem("proccessId") || "",
+      dealId: dealQuery._id || "",
+      docs,
+    });
   };
 
   removeProductItem = (_id) => {
@@ -159,6 +185,62 @@ class ProductForm extends React.Component<Props, State> {
     this.setState({ vatPercent: parseInt(e.currentTarget.value) });
   };
 
+  onChangeCouponCode = e => {
+    this.setState({ couponCode: e.currentTarget.value });
+  };
+
+  applyCoupon = () => {
+    const { dealQuery, productsData, onChangeProductsData, onChangeExtraData } = this.props;
+    const { couponCode } = this.state
+
+    const variables = {
+      _id: dealQuery._id,
+      products: productsData.map(p => ({
+        productId: p.productId,
+        quantity: p.quantity,
+        unitPrice: p.unitPrice
+      })),
+      couponCode
+    };
+
+    client
+      .query({
+        query: gql(queries.checkDiscount),
+        fetchPolicy: "network-only",
+        variables,
+      })
+      .then((res) => {
+        const { checkDiscount } = res.data;
+
+        if (!checkDiscount) {
+          return;
+        }
+
+        const updatedData = (productsData || []).map(p => {
+
+          if (!p.productId || !checkDiscount[p.productId]) {
+            return p;
+          }
+
+          const loyalty = checkDiscount[p.productId]
+
+          const pData = {
+            ...p,
+            discountPercent: loyalty.discount || p.discountPercent || 0
+          };
+    
+          this.calculatePerProductAmount('', pData, false);
+    
+          return pData;
+        })
+
+        onChangeProductsData(updatedData)
+        onChangeExtraData({couponCode})
+
+        this.updateTotal(updatedData);
+      }).catch((err) => Alert.error(err.message));
+  }
+
   applyVat = () => {
     const { productsData, onChangeProductsData } = this.props;
     const { vatPercent } = this.state;
@@ -170,8 +252,8 @@ class ProductForm extends React.Component<Props, State> {
         unitPrice: p.isVatApplied
           ? p.unitPrice
           : parseFloat(
-            ((p.unitPrice * 100) / (100 + (vatPercent || 0))).toFixed(4)
-          ),
+              ((p.unitPrice * 100) / (100 + (vatPercent || 0))).toFixed(4)
+            ),
       };
 
       this.calculatePerProductAmount("", pData, false);
@@ -246,8 +328,13 @@ class ProductForm extends React.Component<Props, State> {
   }
 
   renderContent() {
-    const { productsData, onChangeProductsData, currentProduct, dealQuery, currentUser } =
-      this.props;
+    const {
+      productsData,
+      onChangeProductsData,
+      currentProduct,
+      dealQuery,
+      currentUser,
+    } = this.props;
 
     if (productsData.length === 0) {
       return (
@@ -263,12 +350,10 @@ class ProductForm extends React.Component<Props, State> {
       filteredProductsData = filteredProductsData.filter(
         (p) =>
           p.product &&
-          (
-            p.product.name.includes(filterValues.search) ||
+          (p.product.name.includes(filterValues.search) ||
             p.product.code.includes(filterValues.search) ||
             p.product.shortName.includes(filterValues.search) ||
-            p.product.barcodes.includes(filterValues.search)
-          )
+            p.product.barcodes.includes(filterValues.search))
       );
     }
 
@@ -287,7 +372,8 @@ class ProductForm extends React.Component<Props, State> {
 
     if (filterValues.tags && filterValues.tags.length) {
       filteredProductsData = filteredProductsData.filter(
-        (p) => p.product && lodash.intersection(filterValues.tags, p.product.tagIds)
+        (p) =>
+          p.product && lodash.intersection(filterValues.tags, p.product.tagIds)
       );
     }
 
@@ -546,7 +632,7 @@ class ProductForm extends React.Component<Props, State> {
             name="tags"
             label="Choose tag"
             initialValue={filterValues.tags}
-            onSelect={tagsIds => this.onFilter("tags", tagsIds)}
+            onSelect={(tagsIds) => this.onFilter("tags", tagsIds)}
             multi={true}
           />
         </FormGroup>
@@ -611,13 +697,15 @@ class ProductForm extends React.Component<Props, State> {
     const productOnChange = (products: IProduct[]) => {
       this.clearFilter();
 
-      const { onChangeProductsData, currencies } = this.props;
+      const { onChangeProductsData, currencies, dealsCreateProductData } =
+        this.props;
 
       const { tax, discount } = this.state;
       const currency = currencies ? currencies[0] : "";
 
+      const docs: any[] = [];
       for (const product of products) {
-        productsData.push({
+        const productData = {
           tax: 0,
           taxPercent: tax[currency] ? tax[currency].percent || 0 : 0,
           discount: 0,
@@ -636,8 +724,16 @@ class ProductForm extends React.Component<Props, State> {
           globalUnitPrice: product.unitPrice,
           unitPricePercent: 100,
           _id: Math.random().toString(),
-        });
+        };
+        docs.push(productData);
+        productsData.push(productData);
       }
+
+      dealsCreateProductData({
+        proccessId: localStorage.getItem("proccessId") || "",
+        dealId: dealQuery._id || "",
+        docs,
+      });
 
       onChangeProductsData(productsData);
 
@@ -765,6 +861,27 @@ class ProductForm extends React.Component<Props, State> {
                   </ApplyVatWrapper>
                 </td>
               </tr>
+              {
+                isEnabled('loyalties') &&
+                <tr>
+                <td colSpan={6}>
+                  <ApplyVatWrapper>
+                    <FormControl
+                      placeholder='Coupon code'
+                      onChange={this.onChangeCouponCode}
+                    />
+
+                    <Button
+                      btnStyle='primary'
+                      icon='plus-circle'
+                      onClick={this.applyCoupon}
+                    >
+                      Apply coupon
+                    </Button>
+                  </ApplyVatWrapper>
+                </td>
+              </tr>
+              }
             </tbody>
           </table>
         </FooterInfo>